introducted assertions into sim test singleNodeSingleThread to check for sync committee particiaption
